+++ src/main/java/alisa/AlisaException.java
@@ -1,3 +1,5 @@
+package alisa;
+
 public class AlisaException extends Exception {
     public AlisaException(String message) {
         super(message);
