+++ src/main/java/alisa/Storage.java
@@ -1,3 +1,7 @@
+package alisa;
+
+import alisa.task.*;
+
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileWriter;
