+++ src/main/java/alisa/command/Command.java
@@ -1,3 +1,9 @@
+package alisa.command;
+
+import alisa.AlisaException;
+import alisa.Storage;
+import alisa.task.TaskList;
+import alisa.Ui;
 
 public abstract class Command {
