+++ src/main/java/alisa/command/ExitCommand.java
@@ -1,3 +1,9 @@
+package alisa.command;
+
+import alisa.Storage;
+import alisa.task.TaskList;
+import alisa.Ui;
+
 public class ExitCommand extends Command {
     @Override
     public void execute(TaskList taskList, Ui ui, Storage storage) {
