+++ src/main/java/alisa/command/MarkCommand.java
@@ -1,13 +1,20 @@
-public class DeleteCommand extends Command {
+package alisa.command;
+
+import alisa.AlisaException;
+import alisa.Storage;
+import alisa.task.TaskList;
+import alisa.Ui;
+
+public class MarkCommand extends Command {
     private int index;
 
-    public DeleteCommand(int index) {
+    public MarkCommand(int index) {
         this.index = index;
     }
 
     @Override
     public void execute(TaskList taskList, Ui ui, Storage storage) throws AlisaException {
-        String message = taskList.deleteTask(index);
+        String message = taskList.markTask(index);
         ui.showMessage(message);
         storage.syncFile(taskList);
     }
