+++ src/main/java/alisa/task/Deadline.java
@@ -1,17 +1,19 @@
+package alisa.task;
+
+import alisa.AlisaException;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
 
-public class Event extends Task {
-    private LocalDateTime from;
-    private LocalDateTime to;
+public class Deadline extends Task{
+    private LocalDateTime dueDate;
 
-    public Event(String taskDescription, String from, String to) throws AlisaException {
+    public Deadline(String taskDescription, String dueDate) throws AlisaException {
         super(taskDescription);
         try {
             DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-mm-dd hh:mm");
-            this.from = LocalDateTime.parse(from, formatter);
-            this.to = LocalDateTime.parse(to, formatter);
+            this.dueDate = LocalDateTime.parse(dueDate, formatter);
         } catch (DateTimeParseException e) {
             throw new AlisaException("Please write the deadline in the following format: yyyy-mm-dd hh:mm");
         }
@@ -21,10 +23,10 @@ public class Event extends Task {
     public String toString() {
         String task = super.toString();
         DateTimeFormatter formatter = DateTimeFormatter.ofPattern("MMM dd yyyy hh:mm");
-        return "[E] " + task + " (from: " + from.format(formatter) + " to: " + to.format(formatter) + ")";
+        return "[D] " + task + "(by: " + dueDate.format(formatter) + ")";
     }
     @Override
     public String toFileString() {
-        return "E | " + this.getFileStatus() + " | " + this.getTask() + " | " + from + "-" + to + "\n";
+        return "D | " + this.getFileStatus() + " | " + this.getTask() + " | " + dueDate + "\n";
     }
 }
