+++ src/main/java/alisa/task/Event.java
@@ -1,15 +1,21 @@
+package alisa.task;
+
+import alisa.AlisaException;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
 
-public class Deadline extends Task{
-    private LocalDateTime dueDate;
+public class Event extends Task {
+    private LocalDateTime from;
+    private LocalDateTime to;
 
-    public Deadline(String taskDescription, String dueDate) throws AlisaException {
+    public Event(String taskDescription, String from, String to) throws AlisaException {
         super(taskDescription);
         try {
             DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-mm-dd hh:mm");
-            this.dueDate = LocalDateTime.parse(dueDate, formatter);
+            this.from = LocalDateTime.parse(from, formatter);
+            this.to = LocalDateTime.parse(to, formatter);
         } catch (DateTimeParseException e) {
             throw new AlisaException("Please write the deadline in the following format: yyyy-mm-dd hh:mm");
         }
@@ -19,10 +25,10 @@ public class Deadline extends Task{
     public String toString() {
         String task = super.toString();
         DateTimeFormatter formatter = DateTimeFormatter.ofPattern("MMM dd yyyy hh:mm");
-        return "[D] " + task + "(by: " + dueDate.format(formatter) + ")";
+        return "[E] " + task + " (from: " + from.format(formatter) + " to: " + to.format(formatter) + ")";
     }
     @Override
     public String toFileString() {
-        return "D | " + this.getFileStatus() + " | " + this.getTask() + " | " + dueDate + "\n";
+        return "E | " + this.getFileStatus() + " | " + this.getTask() + " | " + from + "-" + to + "\n";
     }
 }
