+++ src/main/java/alisa/task/Task.java
@@ -1,3 +1,5 @@
+package alisa.task;
+
 public abstract class Task {
     private String taskDescription;
     private boolean isDone;
