+++ src/main/java/alisa/task/Todo.java
@@ -1,3 +1,5 @@
+package alisa.task;
+
 public class Todo extends Task {
 
     public Todo(String taskDescription) {
