+++ text-ui-test/EXPECTED.TXT
@@ -1,4 +1,4 @@
-Hey, Alisa here! What do you need help with?
+Hey, alisa.Alisa here! What do you need help with?
 BTW Say the word bye to get out of this conversation
 ------------------------------------------------------------------------------------
 todo buy groceries
